feat: Implement ensureNotAuthenticated with a kind of state machine
Since the number of paths in logout has lot of possible combinations, we
cannot predict all the possible path and which state of the page takes
place before the next.

I then tried to implement a kind of finite state machine where we
identify all the possible states before the login page :

- connected
- disconnected
- error page
- captcha
- login page
- login page whith only password field
- list of accounts
- keep konnected button
- reload page button
- consent page

For each of these states, we know how to change the state to tend the
final login state and we check the current state in a loop until the
login page state is displayed

If we get an unknown state, an error will be thrown and it will be
possible to know the state with the clisk.html-on-error flag

If the loop is infinit, it will timeout and the logs will display the
path of states.

I hope this will the this connector more stable and more adaptative to
changes in the navigation.